:anchor: Latitude and longitude of the location of the listing is now stored in database

+++ database/migrations/2023_05_13_143845_create_listings_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->string('tags');
             $table->string('company');
             $table->string('location');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->decimal('latitude', 18, 15);
+            $table->decimal('longitude', 18, 15);
             $table->string('email');
             $table->string('website');
             $table->longText('description');
